Extract `Symbol` out into a separate header
