+++ include/caffeine/IR/OperationBase.h
@@ -2,6 +2,7 @@
 
 #include "caffeine/ADT/PersistentArray.h"
 #include "caffeine/ADT/StringInterner.h"
+#include "caffeine/IR/Symbol.h"
 #include "caffeine/IR/Type.h"
 #include "caffeine/Support/Assert.h"
 #include "caffeine/Support/CopyVTable.h"
@@ -81,45 +82,6 @@ enum class FCmpOpcode : uint8_t {
   LE = 004,
   NE = 005,
 };
-
-/**
- * Identifier for a symbolic constant.
- *
- * It can be either a string or a number as required. Numeric symbol names are
- * usually used for internal symbolic values such as allocations. String ones
- * are usually used for user-specified symbolic values.
- */
-class Symbol {
-private:
-  enum {
-    Named = 0,
-    Numbered = 1,
-  };
-
-  std::variant<InternedString, uint64_t> value_;
-
-public:
-  Symbol(const std::string& name);
-  Symbol(std::string&& name);
-  Symbol(std::string_view name);
-  Symbol(uint64_t number);
-
-  template <size_t N>
-  Symbol(const char (&name)[N]);
-
-  bool is_named() const;
-  bool is_numbered() const;
-
-  InternedString name() const;
-  uint64_t number() const;
-
-  bool operator==(const Symbol& symbol) const;
-  bool operator!=(const Symbol& symbol) const;
-
-  friend llvm::hash_code hash_value(const Symbol& symbol);
-};
-
-std::ostream& operator<<(std::ostream& os, const Symbol& symbol);
 
 /**
  * An individual expression node.
@@ -211,8 +173,10 @@ protected:
 protected:
   Operation(std::unique_ptr<OperationData>&& data,
             std::initializer_list<OpRef> operands = {});
-  Operation(Opcode op, Type t, const Inner& inner);
-  Operation(Opcode op, Type t, Inner&& inner);
+  Operation(std::unique_ptr<OperationData>&& data,
+            llvm::ArrayRef<OpRef> operands);
+  [[deprecated]] Operation(Opcode op, Type t, const Inner& inner);
+  [[deprecated]] Operation(Opcode op, Type t, Inner&& inner);
 
   Operation(Opcode op, Type t, const OpRef* operands);
